refactor(desktop): extract CloudSubscriptionManager from WorkspaceGateway
Behavior-preserving decomposition of the 1,826-line WorkspaceGateway god
class (audit-flagged large unit). Extracts the cloud SSE subscription state
machine — the three subscription/retry maps plus subscribe/retry/close logic —
into a dedicated CloudSubscriptionManager, so the gateway keeps thin delegators
with unchanged public signatures.

- new apps/desktop/src/main/cloud-subscription-manager.ts (238 lines): owns
  cloudSessionSubscriptions / cloudWorkspaceSubscriptions / cloudSubscriptionRetryTimers
  and the subscribe/retry-backoff/close-for-workspace logic, constructed with
  resolveWorkspace/getWorkspace/requireCloudAdapter callbacks + reconnect config.

workspace-gateway.ts: 1,826 -> 1,661 lines. Subscribe/retry math and key formats
are byte-identical; the support-matrix block is intentionally left in place
(workspace-authority-contract.test.ts pins those API strings to the gateway file).

Gates: typecheck, eslint --max-warnings 0, knip, node (2238) + workspace (313)
suites, coverage inventory (92.0%) all pass. Public API unchanged.
